Update hasil tangkapan - REMOVE id_jenis_alat_tangkap in hasil_tangkapan table - ADD id_jenis_alat_tangkap & id_jenis_alat_tangkap_lain in detil_hasil_tangkapan table

+++ database/migrations/2016_04_20_133002_create_table_jawaban_konsumsi.php
@@ -17,7 +17,7 @@ class CreateTableJawabanKonsumsi extends Migration
             $table->integer('id_responden');
             $table->integer('id_konsumsi');
             $table->integer('kateg_konsum');
-            $table->float('jawaban')->nullable();
+            $table->double('jawaban')->nullable();
 
             $table->softDeletes();
             $table->timestamps();
